Bug: none
Selenium Harness: move SeleniumServer to a separate class outside of SelNGBase.

1. Create SeleniumService class
2. Change TestBase classes to init the selenium server
3. ZimbraSeleniumProperties ... getIntProperty -> return int instead of long
4. config.properties add seleniumMode=Local (or Remote, Grid, SauceLabs)

+++ ZimbraSelenium/src/java/framework/util/ZimbraSeleniumProperties.java
@@ -29,11 +29,11 @@ public class ZimbraSeleniumProperties {
 		return (getStringProperty(key, null));
 	}
 	
-	public static long getIntProperty(String key) {
+	public static int getIntProperty(String key) {
 		return (getIntProperty(key, 0));
 	}
 
-	public static long getIntProperty(String key, int defaultValue) {
+	public static int getIntProperty(String key, int defaultValue) {
 		String value = ZimbraSeleniumProperties.getInstance().getConfigProp().getString(key, null);
 		if ( value == null )
 			return (defaultValue);
